Add sell tower popup when moving a tower
+ Make inventory GridPane visible in the fxml, but disable on initialization. (So it's visible in SceneBuilder)

+++ src/main/java/seng201/team53/items/towers/QuarryTower.java
@@ -24,8 +24,8 @@ public class QuarryTower extends Tower {
     }
 
     @Override
-    public Double getSellPrice() {
-        return 50.0;
+    public int getSellPrice() {
+        return 50;
     }
 
     @Override
